[refactor] move old entities to a dedicated package
