+++ index.d.ts
@@ -1,5 +1,5 @@
 import ErrorVector from "./src/ErrorVector";
-import IGeoJSONPoint from "./src/IGeoJSONPoint";
-import Path from "./src/Path";
-import Place from "./src/Place";
+import IGeoJSONPoint from "./src/old/IGeoJSONPoint";
+import Path from "./src/old/Path";
+import Place from "./src/old/Place";
 export { ErrorVector, IGeoJSONPoint, Path, Place };
