+++ src/old/IGeoJSONPoint.ts
